ENH: Migrated to new LookingGlass SDK (untested)
lib.exe needs to be on path to build, as the new SDK only has the .dll and .h files and for CMake to work it also needed the .lib and .exp files. To do this I made CMake use lib.exe to create them, after this the new SDK works fine.