ENH: Migrated to new LookingGlass SDK
The Looking Glass Bridge only has the .dll and the .h files, CMake needed also a .lib file. Due to this, a .zip file was created in the releases (https://github.com/xskere/LookingGlassVTKModule/releases/tag/sdk.2.6.0.windows). This file is downloaded and contains all the .dll, .h and .lib needed for Windows.

The .lib file was made using Lib.exe, but I was only able to do it for Windows. For Linux the source code is needed to create the .so file.